Move datetime widget class definitions to components/DateTimeWidget
